+++ common/src/main/java/realcolin/continental/world/continent/ContinentSettings.java
@@ -1,4 +1,4 @@
 package realcolin.continental.world.continent;
 
-public record ContinentSettings(long minContinents, long maxContinents, long meanSize, double variation, double spacing, double uniformity) {
+public record ContinentSettings(int minContinents, int maxContinents, long meanSize, double variation, double spacing, double uniformity) {
 }
